requirements.txt
google-generativeai~=0.4.0
loguru~=0.7.2
neo4j~=5.18.0
numpy~=1.26.4
openai~=1.14.2
python-dotenv~=1.0.1
typer~=0.9.0